Preemptive version bump to 0.3 for Python CI Move Python constants to its own file
