+++ anise-py/src/lib.rs
@@ -19,6 +19,7 @@ use pyo3::prelude::*;
 use pyo3::py_run;
 
 mod astro;
+mod constants;
 mod utils;
 
 /// A Python module implemented in Rust.
